--- src/components/NewButton/types.ts
@@ -1,8 +1,0 @@
-export const CONTAINED_TYPE = 'contained';
-export const OUTLINED_TYPE = 'outlined';
-export const TEXT_TYPE = 'text';
-
-export type ButtonVariantType =
-  | typeof CONTAINED_TYPE
-  | typeof OUTLINED_TYPE
-  | typeof TEXT_TYPE;
